tests/wpt/web-platform-tests/websockets/constructor/001.html

<!doctype html>
<title>WebSockets: new WebSocket() with no args</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
  assert_throws_js(TypeError, function(){new WebSocket()});
});
</script>
